major refactor to move bots into separate folders
