+++ hanabi/bots/__init__.py
@@ -1,0 +1,5 @@
+from .base import BaseBot
+from .dumb import DumbBot
+from .cheating import BasicCheatingBot, CheatingBot
+from .clue import ClueBot, ClueBotImproved, ClueBotAdvanced, ClueBotMk3
+from .lookahead import LookaheadBot
