+++ hanabi/bots/clue.py
@@ -1,81 +1,9 @@
 from __future__ import annotations
 import logging
-
 from operator import and_, not_, or_, eq
 
-from hanabi import Board, CardInfo, Clue, Discard, Play
-
-class BaseBot():
-    def __init__(self):
-        self.reset()
-
-    def reset(self):
-        """Executed once at the start of the game"""
-        pass
-
-    def play(self, board: Board) -> Turn:
-        """Executed on my turn"""
-        raise NotImplemented
-
-
-class DumbBot(BaseBot):
-    """Dumb bot always plays the first card"""
-    def play(self, board: Board) -> Turn:
-        return Play(0)
-
-class BasicCheatingBot(BaseBot):
-    """
-    Cheating bot looks at its hand and plays the last playable cards.
-    Otherwise clues or discards unnecessarily depending on the number of available clues.
-
-    Its main problem is discarding 5s from the initial deal
-    """
-    def play(self, board: Board) -> Turn:
-        hand = board._hands[board.current_player]
-        for i, card in zip(range(board.current_hand_size - 1, -1, -1), reversed(hand)):
-
-            if board.is_playable(card):
-                return Play(i)
-        if board.clues == 0:
-            return Discard(board.current_hand_size - 1)
-        else:
-            return Clue(target=board.relative_player(1), color="r") # Clue the next player red
-
-class CheatingBot(BaseBot):
-    """
-    Improves on BasicCheatingBot by looking at the card it's going to discard and picking safe discards
-    Its weakness is probably getting unlucky with draws and running out of turns
-    """
-    def play(self, board: Board) -> Turn:
-        hand = board._hands[board.current_player]
-
-        # Play any playable cards. Neither ascending or descending order seem to help
-        for i, card in enumerate(hand):
-            if board.is_playable(card):
-                return Play(i)
-
-        # Clue if we're close to max clues
-        if board.clues > 6:
-            return Clue(target=board.relative_player(1), color="r")
-
-        # Discard if we have freely discardable cards
-        for i, card in enumerate(hand):
-            if board.is_discardable(card):
-                return Discard(i)
-
-        # Clue if we have clues left
-        if board.clues > 0:
-            return Clue(target=board.relative_player(1), color="r")
-
-        # Discard any cards that don't need to be saved
-        hand.sort(key=lambda x: x.number, reverse=True) # this discards 4 > 3 > 2, but barely helps
-        for i, card in enumerate(hand):
-            if not board.is_unique(card):
-                return Discard(i)
-
-        # Last resort, discard
-        logging.warning("If this happens, the bot needs to be more sophisticated and take turn order into account")
-        return Discard(0)
+from .base import BaseBot
+from .. import Clue, Discard, Play
 
 class ClueBot(BaseBot):
     """Clues playable cards"""
@@ -305,92 +233,3 @@ class ClueBotAdvanced(BaseBot):
 
         # Couldn't Play. Couldn't Clue. Couldn't Discard. Nothing left to try. This is probably a strike.
         return Play(0)
-
-class LookaheadBot(BaseBot):
-    """
-    LookaheadBot will always try to play a card on its turn. Knowing this about itself, it
-    will assume other players will play on their turns and consider giving clues to cards
-    that will be playable after others have played. 
-
-    For sanity, this 'lookahead' process will only cover each other players next turn, since
-    other players could choose to give us information and we can't predict that. 
-
-    The average score for this bot is 16.38. Slightly better than ClueBotAdvanced.
-    """
-
-    def _is_hypothetical_playable(self, card: Card, played_cards: Dict[str][int]) -> bool:
-        return played_cards[card.color] == card.number-1
-
-    def play(self, board: Board) -> Turn:
-        # If we have a clued card, play it immediately. 
-        for i, card_info in reversed(list(enumerate(board.current_info))):
-            # Play cards from oldest to newest
-            if card_info.clued:
-                return Play(i)
-
-        # Grab a copy of the currently played cards
-        future_cards = {}
-        for color, number in board.played_cards.items():
-            future_cards[color] = number
-
-        # Give a clue, if we can
-        if board.clues > 0:
-            # Decide if a player will play on their turn
-            for player in board.other_players:
-                player_info = board.get_info(player)
-                player_hand = board.get_hand(player)
-
-                will_play = bool([info.clued for info in player_info].count(True))
-
-                if will_play:
-                    # Decide which card they will play, and add it to the hypothetical board state
-                    for i, card_info in reversed(list(enumerate(player_info))):
-                        # Play cards from oldest to newest
-                        if card_info.clued:
-                            # Add this card to the hypothetical board state
-                            card = player_hand[i]
-                            logging.debug(f"Player {player} will play card {i}: {card} on their turn")
-                            if self._is_hypothetical_playable(card, future_cards):
-                                future_cards[card.color] = card.number
-                            else:
-                                logging.debug(f"Player {player} is about to make a mistake!")
-                            break
-                else:
-                    # This player has no cards currently clued. See if we can give them a clue. 
-                    # Follow the rules set out in 'ClueBotAdvanced'
-                    clued_cards = ClueBotAdvanced.get_clued_cards(board)
-                    possible_clues = [Clue(player, number=number) for number in set(board.variant.CARD_NUMBERS)] + [Clue(player, color=color) for color in set(board.variant.CARD_COLORS)]
-
-                    for clue in possible_clues:
-                        clue_touched = board.cards_touched(clue)
-
-                        # A clue must touch at least one card
-                        if sum(clue_touched) == 0:
-                            continue
-
-                        # A clue must not touch any non-(hypothetically)-playable cards:
-                        if sum([(not self._is_hypothetical_playable(card, future_cards)) for idx, card in enumerate(player_hand) if clue_touched[idx]]) > 0:
-                            continue
-
-                        # A clue must not touch a card already clued in another players' hand:
-                        if sum([(card in clued_cards) for idx, card in enumerate(player_hand) if clue_touched[idx]]) > 0:
-                            continue
-
-                        # A clue must not touch multiple of the same card in the same hand:
-                        if max([player_hand.count(card) for idx, card in enumerate(player_hand) if clue_touched[idx]]) > 1:
-                            continue
-
-                        # If we've made it this far, give the clue:
-                        return(clue)
-
-                    logging.debug(f"Player {player} has no valid clues")
-
-        # In this case there are no clues to give. Discard.
-        if board.clues < board.MAX_CLUES:
-            # Oldest card. Any players would have been played already.
-            return Discard(board.current_hand_size-1)
-
-        # Couldn't Play. Couldn't Clue. Couldn't Discard. Nothing left to try. This is probably a strike.
-        return Play(0)
-       
-
